+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ## [Unreleased]
+### Fixed
+- Issue where school CSV import fails if two schools have the same name
 
 ## [4.0.0-alpha6] 2017-06-02
 ### Added